refactor: move get or create table behavior to IcebergTableOperations
The code related to iceberg table "get or create" behavior was
a bit confusing, with some try catches.

This attemps to make it simpler by moving the `loadTable` operation to
IcebergTableOperations and making it return a `Optional<Table>`. To make
us able to use functional-style code, some `Exception` instances were changed to
`RuntimeException`.